Feat: answerHeartClick servarAction으로 변경
