+++ src/libs/actions/answers.ts
@@ -1,0 +1,30 @@
+'use server';
+
+import { cookies } from 'next/headers';
+
+import { apiServer } from '@/utils/apiServer';
+
+import { IClickAnswerHeart } from '../types/response';
+import { reissueAccessToken } from './auth';
+
+export const clickAnswerHeart = async (
+  answerId: number,
+): Promise<IClickAnswerHeart> => {
+  const response = await apiServer.post(`api/v1/answers/${answerId}/hearts`, {
+    cache: 'no-store',
+  });
+
+  if (response.status === 401) {
+    return reissueAccessToken(() => clickAnswerHeart(answerId));
+  }
+
+  if (!response.ok) {
+    const errorData = await response.json();
+    cookies().set('errorMessage', errorData.errorMessage, {
+      maxAge: 5,
+    });
+    throw new Error(`HTTP error! status: ${response.status}`);
+  }
+
+  return response.json();
+};
